Fix notification test and ignore for < API 23
