+++ test/src/main/AndroidManifest.xml
@@ -5,10 +5,11 @@
     <application
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
+        android:name="de.sevenfactory.argon.TestApplication"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".TestActivity"/>
+        <activity android:name="de.sevenfactory.argon.TestActivity"/>
     </application>
 
 </manifest>
