+++ test/src/main/java/de/sevenfactory/argon/MockStore.java
